backend/rag/requirements.txt
chromadb 
openai 
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
pydantic
pyyaml